+++ frontend/components/spot/spot_show/spot_show.js
@@ -51,7 +51,7 @@ class SpotShow extends React.Component {
           starColor="teal"
           emptyStarColor="rgb(215,215,215)"
           // value={this.state.rating_half_star}
-          value={parseInt(spot.averageRating)}
+          value={parseFloat(spot.averageRating)}
           renderStarIcon={(index, value) => {
             return (
               <span>
